refactor(assistant): 提取事件转发器到 forwarder 子模块
