+++ pinvou3-app/src-tauri/src/features/assistant/engine.rs
@@ -85,20 +85,6 @@ struct TurnLifecycleState {
     reclaimed: bool,
     admission_emitted: bool,
     next_reservation_id: u64,
-    /// 单调递增的轮次身份（turn generation）。每次 turn 从 idle 变 active
-    /// 时自增（[`reserve`](TurnLifecycle::reserve) / [`on_submitted`] /
-    /// [`on_started_transition`] 的 newly_active 分支），用于把 cancel 请求
-    /// 绑定到发起时刻的轮次：cancel 在取 `turn_lock` 前后比对 epoch，不匹配
-    /// 说明目标轮已结束（新一轮已 reserve），当前请求必须 no-op，避免跨轮
-    /// 误取消随后启动的新 turn。
-    ///
-    /// 不能复用 `next_reservation_id`：`on_submitted`（定时任务路径）与
-    /// `on_started_transition`（stale TurnStarted）激活 turn 时都不碰它，
-    /// 会让定时 turn 的 generation 恒为旧值。
-    ///
-    /// [`on_submitted`]: TurnLifecycle::on_submitted
-    /// [`on_started_transition`]: TurnLifecycle::on_started_transition
-    turn_epoch: u64,
     active_reservation_id: Option<u64>,
     transcript_rules: Vec<TranscriptSanitizationRule>,
     /// 用户在 turn 已 submit 但尚未收到 TurnStarted 时点了停止。
@@ -113,14 +99,9 @@ struct TurnLifecycleState {
     /// [`take_pending_cancel`] 原子取出并重新 `cancel_current()`——此时
     /// `reset_cancel_token()` 已经执行过，cancel 命中的是当前轮次的活跃 token。
     ///
-    /// 携带 arming 时的 [`turn_epoch`]：并发取消请求（C1/C2）中，排队较晚的
-    /// C2 在恢复后读到的是「当前 lifecycle」（可能已是新轮）。`take_pending_cancel`
-    /// 会校验 epoch 仍是当前轮，跨轮泄漏的 stale pending 被丢弃，不误取消新轮。
-    ///
-    /// [`arm_pending_cancel`]: TurnLifecycle::arm_pending_cancel_and_cancel
+    /// [`arm_pending_cancel`]: TurnLifecycle::arm_pending_cancel
     /// [`take_pending_cancel`]: TurnLifecycle::take_pending_cancel
-    /// [`turn_epoch`]: TurnLifecycleState::turn_epoch
-    pending_cancel: Option<u64>,
+    pending_cancel: bool,
 }
 
 /// The durable, user-visible meaning of a submitted engine operation.
@@ -407,28 +388,6 @@ impl TurnLifecycle {
         state.active || state.terminal_closing
     }
 
-    /// 当前活动轮的 turn generation（epoch），供 cancel 请求绑定发起时刻的
-    /// 轮次身份。与 [`is_active`](Self::is_active) 同口径：turn 活动期间
-    /// （`active` 或终态发送临界区 `terminal_closing`）返回 `Some(epoch)`，
-    /// 空闲返回 `None`。cancel 在取 `turn_lock` 前后各读一次，不匹配则 no-op。
-    pub(crate) fn current_turn_generation(&self) -> Option<u64> {
-        let state = self.state.lock();
-        if state.active || state.terminal_closing {
-            Some(state.turn_epoch)
-        } else {
-            None
-        }
-    }
-
-    /// 当前活动轮是否已提交（`SendMessage` 已入队 / `TurnStarted` 已抵达）。
-    /// 供 cancel 阶段二在 generation mismatch（新轮已 reserve）时判断新轮是否
-    /// 已经真正开始：仅 reserve 未 send 时，engine 里仍是旧轮遗留的子代理，
-    /// 补发级联取消（CancelSubAgents）不会误杀新轮刚启动的子代理（reviewer 点 9）。
-    pub(crate) fn is_current_turn_submitted(&self) -> bool {
-        let state = self.state.lock();
-        state.active && state.submitted
-    }
-
     pub(crate) fn reserve(self: &Arc<Self>) -> Result<TurnReservation> {
         let reservation_id = {
             let mut state = self.state.lock();
@@ -437,7 +396,6 @@ impl TurnLifecycle {
             }
             state.next_reservation_id = state.next_reservation_id.wrapping_add(1).max(1);
             let reservation_id = state.next_reservation_id;
-            state.turn_epoch = state.turn_epoch.wrapping_add(1).max(1);
             state.active = true;
             state.submitted = false;
             state.turn_id = None;
@@ -446,7 +404,7 @@ impl TurnLifecycle {
             state.reclaimed = false;
             state.admission_emitted = false;
             state.active_reservation_id = Some(reservation_id);
-            state.pending_cancel = None;
+            state.pending_cancel = false;
             reservation_id
         };
         Ok(TurnReservation::new(self.clone(), reservation_id))
@@ -627,7 +585,6 @@ impl TurnLifecycle {
     pub(crate) fn on_submitted(&self) -> bool {
         let mut state = self.state.lock();
         if !state.active && !state.terminal_closing {
-            state.turn_epoch = state.turn_epoch.wrapping_add(1).max(1);
             state.active = true;
             state.submitted = true;
             state.turn_id = None;
@@ -662,10 +619,6 @@ impl TurnLifecycle {
         let newly_active = !state.active;
         if newly_active {
             state.admission_emitted = false;
-            // forwarder 收到 stale/repeat TurnStarted 时可能从 idle 激活 turn
-            // （reserve/on_submitted 之外的第三个 active 入口）。同样推进 epoch
-            // 保证 cancel generation 严格单调，覆盖该边缘路径。
-            state.turn_epoch = state.turn_epoch.wrapping_add(1).max(1);
         }
         state.active = true;
         state.submitted = true;
@@ -777,20 +730,6 @@ impl TurnLifecycle {
             .map(|transition| transition.terminal)
     }
 
-    /// 测试专用：认领一个未提交 turn 的终态并立即重开闸门，等价于
-    /// `emit_unsubmitted_interrupted_terminal` 去掉发 `chat:done` 的副作用，
-    /// 供跨模块测试（engine_pool 的 cancel 并发测试）驱动 turn 收尾而无需 AppHandle。
-    /// 不发终态信号——这些测试只关心 lifecycle 状态机的轮次切换，不验证 emit。
-    #[cfg(test)]
-    pub(crate) fn finish_unsubmitted_once(&self) -> bool {
-        if self.claim_unsubmitted_terminal() {
-            self.finish_terminal_emission();
-            true
-        } else {
-            false
-        }
-    }
-
     fn claim_terminal_with_admission(
         &self,
         app: &AppHandle,
@@ -862,27 +801,8 @@ impl TurnLifecycle {
     /// [`claim_terminal_transition`]: Self::claim_terminal_transition
     /// [`claim_reclaimed_transition`]: Self::claim_reclaimed_transition
     pub(crate) fn claim_unsubmitted_terminal(&self) -> bool {
-        self.claim_unsubmitted_terminal_impl(None)
-    }
-
-    /// 带目标 epoch 的原子认领：在 lifecycle state 锁内**同时**校验
-    /// `turn_epoch == target` 并认领，二者之间不存在可插入轮次切换的窗口。
-    ///
-    /// 供 cancel 阶段二使用（`reserve_turn` 不取 `turn_lock`，generation 检查
-    /// 与认领若分离，另一 worker 可在检查通过后结束旧轮并 reserve 新轮，
-    /// 无 epoch 校验的认领会把新轮 reservation 误认领为 Interrupted）：
-    /// epoch 不匹配时返回 `false` 且不产生任何副作用，调用方必须整体 no-op。
-    pub(crate) fn claim_unsubmitted_terminal_for_epoch(&self, target: u64) -> bool {
-        self.claim_unsubmitted_terminal_impl(Some(target))
-    }
-
-    fn claim_unsubmitted_terminal_impl(&self, expected_epoch: Option<u64>) -> bool {
         let _emission = self.emission.lock();
         let mut state = self.state.lock();
-        if expected_epoch.is_some_and(|epoch| state.turn_epoch != epoch) {
-            // 轮次已切换（目标轮已结束、新轮已 reserve）：不得认领新轮。
-            return false;
-        }
         if !state.active || state.submitted || state.terminal_emitted {
             return false;
         }
@@ -920,28 +840,10 @@ impl TurnLifecycle {
         true
     }
 
-    /// 带目标 epoch 的未提交认领 + 补发 `chat:done`：认领在 state 锁内与
-    /// `turn_epoch == target` 校验原子完成，目标轮已结束（新轮已 reserve）时
-    /// 返回 `false` 且无副作用，避免把新轮 reservation 误认领为 Interrupted。
-    pub(crate) fn emit_unsubmitted_interrupted_terminal_for_epoch(
-        &self,
-        app: &AppHandle,
-        session_id: &str,
-        target: u64,
-    ) -> bool {
-        if !self.claim_unsubmitted_terminal_for_epoch(target) {
-            return false;
-        }
-        emit_chat_terminal(app, session_id, TurnOutcomeStatus::Interrupted, None, false);
-        self.finish_terminal_emission();
-        true
-    }
-
     /// 标记「cancel 在 turn 已 submit 但尚未 TurnStarted 时发起」。
     ///
     /// 仅当 turn 处于 active、已 `submitted`、且 `turn_id` 仍为 None（TurnStarted
-    /// 未抵达）、且 `turn_epoch == epoch`（仍是发起 cancel 的那一轮）时设置
-    /// `pending_cancel = Some(epoch)`：
+    /// 未抵达）时设置 `pending_cancel`：
     /// - 必须 `submitted`：未提交的 reservation（消息尚未入队 engine）应由 cancel
     ///   走未提交认领终态路径（`emit_unsubmitted_interrupted_terminal`）立即发
     ///   `chat:done` 使 reservation 失效，而不是挂成 pending——否则空闲 engine 仍
@@ -949,88 +851,16 @@ impl TurnLifecycle {
     ///   前端 busy 在 cancel 后到 TurnStarted 之间无法复位。
     /// - 若 `turn_id` 已有值说明 TurnStarted 已被转发器消费，`cancel_current()`
     ///   直接命中当前活跃 token，无需补打。
-    /// - 必须 `turn_epoch == epoch`：并发取消请求（C1/C2）中，排队较晚的 C2 在
-    ///   持锁恢复后读到的是「当前 lifecycle」。cancel 已在取 `turn_lock` 前后比对
-    ///   过 epoch（见 [`current_turn_generation`]/cancel 路径），此处传入**当前**
-    ///   epoch 作二次锚定，确保 pending 只 arm 到目标轮，不跨轮泄漏到新轮的
-    ///   TurnStarted（forwarder 经 [`take_pending_cancel`] 校验 epoch 后才重放）。
     ///
     /// **调用顺序**：必须在 `cancel_current()` **之前**调用。两者取不同的锁
     /// （lifecycle state mutex vs cancel_token mutex），无法原子合并。先 arm
     /// 再 cancel 保证：即使 TurnStarted 在两步之间抵达转发器并消费了标记，
     /// 随后的 `cancel_current()` 也只是幂等 no-op（转发器已重新 cancel）。
-    ///
-    /// **返回值**：`false` 表示 `state.turn_epoch != epoch`——generation 复查
-    /// 通过之后、arm 之前另一 worker 已结束目标轮并启动新轮（`reserve_turn`
-    /// 不取 `turn_lock`，可在 cancel 的同步段中间完成切换）。调用方**必须**
-    /// 在收到 `false` 时跳过 `cancel_current` 及其级联副作用，否则取消会命中
-    /// 新轮已 `reset_cancel_token` 的活跃 token，造成跨轮误取消。
-    ///
-    /// 其余条件不满足（未 submitted / 已 started / 非 active）时仍返回 `true`：
-    /// 这些情况 epoch 匹配、仍是目标轮，`cancel_current` 命中目标轮 token
-    /// （已 started 直接命中活跃 token；未 submitted 时 engine 尚无该轮 token，
-    /// 取消是幂等 no-op），调用方可以安全继续。
-    ///
-    /// [`current_turn_generation`]: Self::current_turn_generation
-    /// [`take_pending_cancel`]: Self::take_pending_cancel
-    ///
-    /// 在 lifecycle state 锁内原子完成「epoch 校验 + arm pending + 同步取消」。
-    ///
-    /// 与 [`arm_pending_cancel`] 的区别：`cancel` 闭包在**同一临界区内**持锁
-    /// 执行。单独 arm 时锁在校验后即释放，调用方随后才执行 `cancel_current`，
-    /// 两条同步调用之间没有 `.await` 也不构成原子性保证——多线程 runtime/OS
-    /// 可以在任意指令边界切换线程。若另一 worker 在该窗口内完成「旧轮终态
-    /// 收口 + 新轮 reserve/send + `reset_cancel_token`」，恢复后的旧 cancel 会
-    /// 命中新轮活跃 token（reviewer 点 8）。把取消闭包移入同一临界区后，
-    /// `reserve_turn`（取同一把 state 锁）无法插入「校验/arm」与「取消」之间，
-    /// 跨轮窗口闭合。
-    ///
-    /// **idle 守卫**（G2/epoch-0 哨兵）：`turn_epoch` 从 1 起自增，但 fresh
-    /// 会话（从未有过 turn）初始为 0。此时发起 cancel 快照 `target = None`，
-    /// 调用方以 `target.unwrap_or(0)` 传入后，`turn_epoch != epoch` 校验（0 != 0
-    /// 不成立）会通过——若 engine 已为「即将启动的定时轮」在场（`run_scheduled_turn`
-    /// 的 spawn→submit 窗口），执行 `cancel_current` 会命中该新 token，取消信号
-    /// 与 `reset_cancel_token` 竞速、结果不确定（定时轮被意外取消或停止丢失）。
-    /// 因此 epoch 匹配后仍需 `state.active || state.terminal_closing` 才执行
-    /// `cancel` 闭包：空闲（无活跃轮）时跳过取消（幂等 no-op），但返回 `true`
-    /// 让调用方继续（级联取消仍由调用方按 armed 路径执行，只取消 engine 遗留
-    /// 子代理、不取消尚未启动的轮）。
-    ///
-    /// 锁序：本方法持 lifecycle state 锁调用 `cancel` 闭包，闭包内
-    /// `engine.cancel_current()` 取 engine 的 cancel_token 锁（与 state 锁无
-    /// 反向依赖，forwarder 消费 pending 也是先 state 后 token），无死锁。
-    /// `cancel` 必须同步、不 panic（panic 会使 Mutex 中毒），且不得再次获取
-    /// 本 lifecycle 的 state 锁。
-    ///
-    /// 其余语义与 [`arm_pending_cancel`] 一致：epoch 匹配则设置 pending（条件
-    /// 满足时）并执行 `cancel` 后返回 `true`；epoch 不匹配返回 `false` 且
-    /// **不执行** `cancel`，调用方必须整体 no-op。
-    ///
-    /// [`arm_pending_cancel`]: Self::arm_pending_cancel_and_cancel
-    pub(crate) fn arm_pending_cancel_and_cancel<F>(&self, epoch: u64, cancel: F) -> bool
-    where
-        F: FnOnce(),
-    {
+    pub(crate) fn arm_pending_cancel(&self) {
         let mut state = self.state.lock();
-        if state.turn_epoch != epoch {
-            // 轮次已切换（目标轮已结束、新轮已 reserve）：不 arm、不取消。
-            return false;
-        }
-        if !state.active && !state.terminal_closing {
-            // idle（无活跃轮、非终态收口期）：target 快照为 None 时以 0 哨兵
-            // 传入，epoch 校验通过但并无可取消的轮。执行 cancel 会命中「即将
-            // 启动的定时轮」的新 token（G2）。跳过取消闭包、不 arm（idle 也
-            // 不满足 submitted 前置），但返回 true：调用方按 armed 路径继续，
-            // 级联取消只清 engine 遗留子代理，不误伤尚未启动的轮。
-            return true;
-        }
         if state.submitted && state.turn_id.is_none() && state.active {
-            state.pending_cancel = Some(epoch);
+            state.pending_cancel = true;
         }
-        // 持锁执行同步取消：reserve_turn 需要同一把 state 锁，无法在
-        // 「校验/arm」与「取消」之间插入轮次切换，旧 cancel 不可能命中新轮。
-        cancel();
-        true
     }
 
     /// 原子取出并清除 `pending_cancel` 标记。
@@ -1038,20 +868,9 @@ impl TurnLifecycle {
     /// 由事件转发器在收到 `TurnStarted` 后调用：此时 CodeWhale 的
     /// `reset_cancel_token()` 已执行完毕（它在 `TurnStarted` 之前），
     /// 重新 `cancel_current()` 命中的正是本轮的活跃 token。
-    ///
-    /// 仅当记录的 epoch 仍是 `current_epoch`（仍是 arming 时的那一轮）时才取出
-    /// 并返回 `Some`，否则清空并返回 `None`：跨轮泄漏的 stale pending（cancel
-    /// arm 到旧轮后，新一轮 TurnStarted 先抵达）被丢弃，不误取消新轮。空闲时
-    /// `current_epoch` 由调用方传 0（epoch 自增从 1 起，恒不匹配）。
-    pub(crate) fn take_pending_cancel(&self, current_epoch: u64) -> Option<u64> {
+    pub(crate) fn take_pending_cancel(&self) -> bool {
         let mut state = self.state.lock();
-        match state.pending_cancel {
-            Some(epoch) if epoch == current_epoch => state.pending_cancel.take(),
-            _ => {
-                state.pending_cancel = None;
-                None
-            }
-        }
+        std::mem::take(&mut state.pending_cancel)
     }
 }
 
@@ -1915,1404 +1734,9 @@ pub(crate) async fn apply_harness_action(
     }
 }
 
-/// 后台 task：持续读 rx_event 转 Tauri emit。
-///
-/// 关键点：监听 `Event::ApprovalRequired` 并主动 `approve_tool_call`。
-/// 注意底座语义：`auto_approve = true` 时 `registered_tool_approval_required`
-/// 会直接**旁路**可绕过的 Required 审批、不发事件（turn_loop.rs，非绕过名单仅
-/// rlm_eval / start_mcp_server）。因此本臂对普通会话只在 auto_approve 关闭的
-/// 场景（定时任务按 profile、工作流运行按 `apply_workflow_turn_policy`）收到
-/// 事件；工作流编排图交给用户裁定，工作流宿主的其它副作用工具拒绝，普通会话
-/// 仍沿用既有批准策略。
-///
-/// Plan ready 触发：监听 `update_plan` 工具结果 + TurnComplete，
-/// 若 active session mode=Plan + 本 turn 调过 update_plan + plan 非空 →
-/// 设 phase=Ready + emit `chat:plan_ready` 含 plan snapshot。
-///
-/// **多引擎并发**:每个 session 一个独立 engine + 一个独立 forwarder,所以这里捕获
-/// 的 `session_id` 唯一标识本 forwarder 服务的 session。**所有 emit 的 payload 都带
-/// `session_id`**,前端按它把事件分流到对应 session 的缓冲;TurnComplete 里的 mode
-/// 判据(plan_ready / M2 / M3)也全部基于本 `session_id`,不再读全局 `store.active_id()`
-/// (并发下 active 会变,读全局会把判据算到错误 session 上)。返回 forwarder 的
-/// `JoinHandle`,EnginePool 回收 session 时 `abort()` 它。
-fn spawn_event_forwarder(
-    app: AppHandle,
-    handle: EngineHandle,
-    store: SessionStore,
-    bridge: Pinvou3Bridge,
-    execution_workspace: PathBuf,
-    session_id: String,
-    turn_events: broadcast::Sender<EngineTurnSignal>,
-    scheduled_profile: Option<ScheduledRunProfile>,
-    scheduled_base_total_tokens: Option<u64>,
-    scheduled_unattended: Arc<AtomicBool>,
-    turn_lifecycle: Arc<TurnLifecycle>,
-    shell_manager: SharedShellManager,
-    turn_shell_tasks: TurnShellTaskRegistry,
-) -> tauri::async_runtime::JoinHandle<()> {
-    let approve_handle = handle.clone();
-    let plan_tracker: Arc<Mutex<TurnPlanTracker>> =
-        Arc::new(Mutex::new(TurnPlanTracker::default()));
-    let shell_output = crate::features::assistant::shell_output::ShellOutputMonitor::spawn(
-        app.clone(),
-        &shell_manager,
-        session_id.clone(),
-    );
-    tauri::async_runtime::spawn(async move {
-        // [B2] 完成账本:agent_id -> 决策 role。dedup——同一 agent_id 重复
-        // AgentComplete 时跳过推进(防双推进)。角色重派(gate 失败/回滚)得新
-        // agent_id,不会被误跳。
-        let mut seen_completions: std::collections::HashMap<String, String> =
-            std::collections::HashMap::new();
-        // [edict-obs] agent_id→role_id 关联(由 fork 的 AgentSpawned 第二发喂,
-        // 见下方 AgentSpawned 臂注释)+ per-role token 账本。
-        // 有意不清理:存活期=本 session forwarder,单 run 最多几十条目(含 fan-out 重试),
-        // 跟 seen_completions 同模式 —— 别加"AgentComplete 时清理",会破坏 dedup 语义。
-        let mut agent_roles: std::collections::HashMap<String, String> =
-            std::collections::HashMap::new();
-        let mut tool_inputs: std::collections::HashMap<String, (String, serde_json::Value)> =
-            std::collections::HashMap::new();
-        let mut token_ledger = TokenLedger::default();
-        let mut turn_tracker = TurnCompletionTracker::default();
-        let mut scheduled_engine_total_tokens = 0_u64;
-        let mut scheduled_persistence_error: Option<String> = None;
-        let mut latest_chat_engine_state: Option<ChatEngineState> = None;
-        let mut chat_persistence_error: Option<String> = None;
-        let mut active_transcript_seen = false;
-        // app 侧自测推理指标累加器(TTFT/生成速度/累计tokens/KV)。try_state:headless
-        // harness / 测试可能没 manage MonitorState,拿不到就整块跳过,不 panic。
-        let self_metrics = app
-            .try_state::<crate::features::monitor::MonitorState>()
-            .map(|s| s.self_metrics());
-        let mut current_turn_id: Option<String> = None;
-        let mut rx = handle.rx_event.write().await;
-        while let Some(event) = rx.recv().await {
-            match event {
-                Event::TurnStarted { turn_id, .. } => {
-                    // Publish admission from the authoritative engine event,
-                    // before this serial forwarder can observe any delta or
-                    // terminal event for the same turn. Reclaim uses the same
-                    // emission gate, so it cannot overtake this pair.
-                    let admitted =
-                        turn_lifecycle.emit_started_admission(&app, &session_id, turn_id.clone());
-                    // 消费 pending_cancel（无论 admitted 与否，防止跨轮泄漏）。
-                    // reset_cancel_token() 在 TurnStarted 之前已执行，若 cancel
-                    // 在此之前 arm 了标记，现在重新 cancel 命中的是本轮活跃 token。
-                    //
-                    // 仅当 pending_cancel 记录的 epoch 仍是当前轮（emit_started_admission
-                    // 刚为该轮推进了 epoch）才取出重放；并发取消请求在旧轮 arm 的
-                    // stale pending 会被丢弃，不误取消新轮。
-                    let current_epoch = turn_lifecycle.current_turn_generation().unwrap_or(0);
-                    let pending_cancel = turn_lifecycle.take_pending_cancel(current_epoch);
-                    if !admitted {
-                        continue;
-                    }
-                    if pending_cancel.is_some() {
-                        approve_handle.cancel();
-                    }
-                    active_transcript_seen = false;
-                    chat_persistence_error = None;
-                    current_turn_id = Some(turn_id.clone());
-                    if let Err(error) = turn_shell_tasks.bind_or_prepare_turn(&turn_id).await {
-                        log::error!(
-                            "[pinvou3][chat] failed to bind shell task scope sid={} turn={}: {error:#}",
-                            session_id,
-                            turn_id
-                        );
-                    }
-                    let _ = turn_events.send(turn_tracker.on_started(turn_id));
-                    // 本轮起始打点(TTFT 起点)。底座已发此事件,原先落 `_` 被忽略。
-                    if let Some(m) = &self_metrics {
-                        m.on_turn_started(&session_id);
-                    }
-                }
-                Event::MessageDelta { content, .. } => {
-                    if let Some(m) = &self_metrics {
-                        m.on_message_delta(&session_id, content.chars().count());
-                    }
-                    crate::features::memory::append_turn_assistant(&session_id, &content);
-                    let payload = json!({ "session_id": session_id, "text": content });
-                    let _ = app.emit("chat:delta", payload.clone());
-                    crate::features::remote_control::forward_app_event(&app, "chat:delta", payload);
-                }
-                Event::ThinkingStarted { index } => {
-                    let payload = json!({ "session_id": session_id, "index": index });
-                    let _ = app.emit("chat:reasoning_start", payload.clone());
-                    crate::features::remote_control::forward_app_event(
-                        &app,
-                        "chat:reasoning_start",
-                        payload,
-                    );
-                }
-                Event::ThinkingDelta { index, content } => {
-                    // 只转发底座已经识别为 ThinkingDelta 的独立思考内容。普通
-                    // MessageDelta 不做启发式猜测，避免把最终回答误折叠成思考。
-                    let payload =
-                        json!({ "session_id": session_id, "index": index, "text": content });
-                    let _ = app.emit("chat:reasoning_delta", payload.clone());
-                    crate::features::remote_control::forward_app_event(
-                        &app,
-                        "chat:reasoning_delta",
-                        payload,
-                    );
-                }
-                Event::ThinkingComplete { index } => {
-                    let payload = json!({ "session_id": session_id, "index": index });
-                    let _ = app.emit("chat:reasoning_done", payload.clone());
-                    crate::features::remote_control::forward_app_event(
-                        &app,
-                        "chat:reasoning_done",
-                        payload,
-                    );
-                }
-                Event::ToolCallStarted { id, name, input } => {
-                    if let Some(m) = &self_metrics {
-                        m.on_tool(&session_id); // 本轮有工具 → 收尾跳过 TTFT/TPS(D2)
-                    }
-                    crate::features::memory::record_turn_tool_start(&session_id, &name, &input);
-                    shell_output.tool_started(&id, &name, &input);
-                    tool_inputs.insert(id.clone(), (name.clone(), input.clone()));
-                    let payload =
-                        json!({ "session_id": session_id, "id": id, "name": name, "args": input });
-                    let _ = app.emit("chat:tool_start", payload.clone());
-                    crate::features::remote_control::forward_app_event(
-                        &app,
-                        "chat:tool_start",
-                        payload,
-                    );
-                }
-                Event::ToolCallComplete { id, name, result } => {
-                    // 携带 metadata 让前端识别 careful hook 拦截 (safety_level=="dangerous")
-                    let (output, success, metadata) = tool_call_result_parts(result);
-                    let background_task_id =
-                        if matches!(name.as_str(), "exec_shell" | "task_shell_start" | "Bash")
-                            && metadata
-                                .as_ref()
-                                .and_then(|value| value.get("status"))
-                                .and_then(serde_json::Value::as_str)
-                                == Some("Running")
-                        {
-                            metadata
-                                .as_ref()
-                                .and_then(|value| value.get("task_id"))
-                                .and_then(serde_json::Value::as_str)
-                                .map(str::to_owned)
-                        } else {
-                            None
-                        };
-                    if let (Some(turn_id), Some(task_id)) =
-                        (current_turn_id.as_deref(), background_task_id.as_deref())
-                    {
-                        if !turn_shell_tasks.register_task(turn_id, task_id) {
-                            log::warn!(
-                                "[pinvou3][chat] shell task has no root-turn scope sid={} turn={} task={}",
-                                session_id,
-                                turn_id,
-                                task_id
-                            );
-                        }
-                    }
-                    shell_output.tool_completed(&id, background_task_id.as_deref());
-                    let tracked_input = tool_inputs
-                        .remove(&id)
-                        .map(|(_, input)| input)
-                        .unwrap_or(serde_json::Value::Null);
-                    if success {
-                        if let Some(profile) = scheduled_profile.as_ref() {
-                            let artifact_store = store.clone();
-                            let artifact_session_id = session_id.clone();
-                            let artifact_workspace = profile.workspace.clone();
-                            let artifact_name = name.clone();
-                            let artifact_input = tracked_input.clone();
-                            let artifact_output = output.clone();
-                            match tokio::task::spawn_blocking(move || {
-                                persist_successful_tool_artifact(
-                                    &artifact_store,
-                                    &artifact_session_id,
-                                    &artifact_workspace,
-                                    &artifact_name,
-                                    &artifact_input,
-                                    &artifact_output,
-                                )
-                            })
-                            .await
-                            {
-                                Ok(Ok(_)) => {}
-                                Ok(Err(error)) => eprintln!(
-                                    "[pinvou3-app] scheduled artifact persistence skipped: {error:#}"
-                                ),
-                                Err(error) => eprintln!(
-                                    "[pinvou3-app] scheduled artifact persistence task failed: {error}"
-                                ),
-                            }
-                        }
-                    }
-                    crate::features::memory::record_turn_tool_complete(
-                        &session_id,
-                        &name,
-                        &tracked_input,
-                        success,
-                    );
-                    // Plan 类工具结果：标记 + 缓存 snapshot（两层）+ 实时 emit 给前端 chip 进度区
-                    if success
-                        && (name == "update_plan"
-                            || name == "checklist_write"
-                            || name == "todo_write")
-                    {
-                        let mut tracker = plan_tracker.lock();
-                        tracker.plan_tool_used = true;
-                        // 上游格式："Plan/Todo ... updated: ...\n{json}"——切第一个 '\n' 后是 json
-                        if let Some(json_part) = output.find('\n').map(|i| &output[i + 1..]) {
-                            if let Ok(v) = serde_json::from_str::<serde_json::Value>(json_part) {
-                                if name == "update_plan" {
-                                    tracker.last_plan_snapshot = Some(v);
-                                } else {
-                                    tracker.last_todos_snapshot = Some(v);
-                                }
-                            }
-                        }
-                        // emit chat:plan_snapshot 实时更新 chip 进度——跟 plan_ready 解耦,
-                        // 后者是状态转移信号(Planning→Ready 弹卡片),这个是数据更新信号(刷新 chip)。
-                        // **只 emit 本次工具改的那个 snapshot**,另一个置 null——这样前端只更新对应
-                        // 的时间戳,pickProgressItems 才能正确按时间挑最新的(否则两个 ts 总相等)。
-                        let (plan_emit, todos_emit) = if name == "update_plan" {
-                            (tracker.last_plan_snapshot.clone(), None)
-                        } else {
-                            (None, tracker.last_todos_snapshot.clone())
-                        };
-                        drop(tracker);
-                        let payload = json!({
-                            "session_id": session_id,
-                            "plan_snapshot": plan_emit,
-                            "todos_snapshot": todos_emit,
-                        });
-                        let _ = app.emit("chat:plan_snapshot", payload.clone());
-                        crate::features::remote_control::forward_app_event(
-                            &app,
-                            "chat:plan_snapshot",
-                            payload,
-                        );
-                    }
-                    let payload = json!({
-                        "session_id": session_id,
-                        "id": id,
-                        "name": name,
-                        "output": output,
-                        "success": success,
-                        "metadata": metadata,
-                    });
-                    // request_user_input 收口：submit/cancel 已由底座转为工具结果。
-                    if name == "request_user_input" {
-                        crate::features::assistant::pending_user_input::clear(&session_id, &id);
-                    }
-                    let _ = app.emit("chat:tool_end", payload.clone());
-                    crate::features::remote_control::forward_app_event(
-                        &app,
-                        "chat:tool_end",
-                        payload,
-                    );
-                }
-                Event::UserInputRequired { id, request } => {
-                    if scheduled_profile.is_some() && scheduled_unattended.load(Ordering::Acquire) {
-                        // 定时任务没有前台操作者；不能把底座卡在等待输入的超时上。
-                        let h = approve_handle.clone();
-                        tokio::spawn(async move {
-                            if let Err(error) = h.cancel_user_input(id).await {
-                                eprintln!(
-                                    "[pinvou3-app] cancel scheduled user input failed: {error:?}"
-                                );
-                            }
-                        });
-                    } else {
-                        // 普通对话继续交给前端或远程控制端处理。登记 pending，
-                        // 供前端 remount 后经 get_pending_user_inputs 恢复确认卡。
-                        let payload = json!({
-                            "session_id": session_id,
-                            "id": id,
-                            "questions": request.questions,
-                        });
-                        crate::features::assistant::pending_user_input::record(
-                            &session_id,
-                            &id,
-                            payload["questions"].clone(),
-                        );
-                        let _ = app.emit("chat:user_input_required", payload.clone());
-                        crate::features::remote_control::forward_app_event(
-                            &app,
-                            "chat:user_input_required",
-                            payload,
-                        );
-                    }
-                }
-                Event::SessionUpdated {
-                    session_id: event_session_id,
-                    messages,
-                    system_prompt,
-                    model,
-                    workspace,
-                } => {
-                    if event_session_id != session_id {
-                        let mismatch = format!(
-                            "Engine session id mismatch: expected {session_id}, got {event_session_id}"
-                        );
-                        if scheduled_profile.is_some() {
-                            scheduled_persistence_error = Some(mismatch);
-                        } else {
-                            chat_persistence_error = Some(mismatch);
-                        }
-                        continue;
-                    }
-                    if let Some(profile) = scheduled_profile.as_ref() {
-                        let state = ScheduledEngineState {
-                            messages,
-                            system_prompt,
-                            model,
-                            workspace,
-                            mode: profile.execution_mode(),
-                            token_accounting: ScheduledTokenAccounting::PreservePersisted,
-                        };
-                        let store_for_save = store.clone();
-                        let session_for_save = session_id.clone();
-                        match tokio::task::spawn_blocking(move || {
-                            store_for_save.persist_scheduled_engine_state(&session_for_save, state)
-                        })
-                        .await
-                        {
-                            Ok(Ok(_)) => scheduled_persistence_error = None,
-                            Ok(Err(error)) => {
-                                scheduled_persistence_error = Some(format!("{error:#}"));
-                            }
-                            Err(error) => {
-                                scheduled_persistence_error = Some(format!(
-                                    "scheduled transcript persistence task failed: {error}"
-                                ));
-                            }
-                        }
-                    } else {
-                        let (messages, matched_active_rule) =
-                            turn_lifecycle.sanitize_messages(messages);
-                        active_transcript_seen |= matched_active_rule;
-                        let state = ChatEngineState {
-                            messages,
-                            system_prompt,
-                            model,
-                            workspace,
-                        };
-                        latest_chat_engine_state = Some(state.clone());
-                        let store_for_save = store.clone();
-                        let session_for_save = session_id.clone();
-                        match tokio::task::spawn_blocking(move || {
-                            store_for_save.persist_chat_engine_state(&session_for_save, state)
-                        })
-                        .await
-                        {
-                            Ok(Ok(saved)) => {
-                                chat_persistence_error = None;
-                                if let Ok(revision) = transcript_revision(&saved.messages) {
-                                    let payload = json!({
-                                        "session_id": session_id,
-                                        "transcript_revision": revision,
-                                    });
-                                    let _ = app.emit("chat:transcript_committed", payload.clone());
-                                    crate::features::remote_control::forward_app_event(
-                                        &app,
-                                        "chat:transcript_committed",
-                                        payload,
-                                    );
-                                }
-                            }
-                            Ok(Err(error)) => {
-                                chat_persistence_error = Some(format!("{error:#}"));
-                            }
-                            Err(error) => {
-                                chat_persistence_error = Some(format!(
-                                    "chat transcript persistence task failed: {error}"
-                                ));
-                            }
-                        }
-                    }
-                }
-                Event::ApprovalRequired {
-                    id,
-                    tool_name,
-                    approval_force_prompt,
-                    ..
-                } => {
-                    // 多智能体会话与普通对话共用同一套审批语义（ADR-0006）：
-                    // 只有无人值守的初始定时执行使用任务保存的批准策略；
-                    // 其余会话按普通对话处理。
-                    let should_approve = if scheduled_profile.is_some()
-                        && scheduled_unattended.load(Ordering::Acquire)
-                    {
-                        scheduled_tool_should_auto_approve(
-                            scheduled_profile.as_ref(),
-                            approval_force_prompt,
-                        )
-                    } else {
-                        true
-                    };
-                    eprintln!(
-                        "[pinvou3-app] {} tool {} id={}",
-                        if should_approve {
-                            "auto-approving"
-                        } else {
-                            "denying"
-                        },
-                        tool_name,
-                        id
-                    );
-                    let h = approve_handle.clone();
-                    let id_clone = id.clone();
-                    tokio::spawn(async move {
-                        let result = if should_approve {
-                            h.approve_tool_call(id_clone).await
-                        } else {
-                            h.deny_tool_call(id_clone).await
-                        };
-                        if let Err(e) = result {
-                            eprintln!("[pinvou3-app] scheduled approval decision failed: {e:?}");
-                        }
-                    });
-                    // 不重复 emit chat:tool_start —— 上游 ToolCallStarted（带完整 input）
-                    // 已先于 ApprovalRequired fire，前端已收到正确的 args。
-                    // 之前在此 emit 会用 args=null 覆盖前端 toolMeta，导致产物路径丢失。
-                }
-                // [edict-obs] 同一 agent_id 会收到两发 AgentSpawned:第一发来自
-                // subagent manager 内部(prompt=任务文本,subagent/mod.rs:1518),
-                // 第二发来自 fork 的 Op::SpawnSubAgent 成功臂(prompt=role_id)。
-                // FIFO 保证第二发后到 → HashMap::insert last-write-wins,最终值
-                // 必为 role_id。这是有意设计,别"去重优化"。
-                // v0.8.65 上游给 AgentSpawned 加 parent_run_id/spawn_depth(谱系遥测);
-                // pinvou3 forwarder 只用 id→role(prompt)关联,新字段 `..` 忽略。
-                Event::AgentSpawned { id, prompt, .. } => {
-                    agent_roles.insert(id, prompt);
-                }
-                // 台账退役（ADR-0006）：workflow 内部生命周期不再转发，子智能体
-                // 状态一律走 AgentProgress / AgentComplete 与落盘 worker ledger。
-                Event::WorkflowUi { .. } => {}
-                // [edict-obs] SubAgent 每步进展(底座自动发,不靠 prompt 纪律)→ 前端看板。
-                Event::AgentProgress { id, status, .. } => {
-                    // 早期事件可能赶在第二发 AgentSpawned(role_id)之前到 —— 兜底用
-                    // agent_id 而不是空串,跟 TokenUsage 臂的 fallback 语义一致。
-                    let role = agent_roles.get(&id).cloned().unwrap_or_else(|| id.clone());
-                    let payload = json!({
-                        "session_id": session_id,
-                        "agent_id": id,
-                        "role_id": role,
-                        "status": status,
-                    });
-                    let _ = app.emit("workflow:agent_progress", payload);
-                }
-                // [edict-obs] mailbox 信封:工具调用→progress;TokenUsage→账本+审计;
-                // Completed/Failed→审计。审计写盘是同步 IO,但单行 append 微秒级,
-                // 不值得为它 spawn_blocking(forwarder 本身不在 LLM 关键路径上)。
-                Event::SubAgentMailbox { message, .. } => {
-                    use deepseek_tui::tools::subagent::MailboxMessage as MM;
-                    // 审计是应用账本：绑了项目目录的原生代码会话写会话私有目录，
-                    // 不污染用户项目；其余会话账本根与执行根相同，行为不变。
-                    let ws = bridge.audit_workspace(&session_id, &execution_workspace);
-                    match message {
-                        MM::Started { agent_id, .. } => {
-                            if let Some(turn_id) = current_turn_id.as_deref() {
-                                turn_shell_tasks.register_agent(turn_id, &agent_id);
-                            }
-                        }
-                        MM::ChildSpawned {
-                            parent_id,
-                            child_id,
-                        } => {
-                            if !turn_shell_tasks.register_child_agent(&parent_id, &child_id) {
-                                log::warn!(
-                                    "[pinvou3][chat] child agent has no parent shell scope sid={} parent={} child={}",
-                                    session_id,
-                                    parent_id,
-                                    child_id
-                                );
-                            }
-                        }
-                        MM::TokenUsage {
-                            agent_id,
-                            route,
-                            usage,
-                            ..
-                        } => {
-                            let role = agent_roles
-                                .get(&agent_id)
-                                .cloned()
-                                .unwrap_or_else(|| agent_id.clone());
-                            let (input_total, output_total, calls) = token_ledger.add(
-                                &role,
-                                usage.input_tokens as u64,
-                                usage.output_tokens as u64,
-                            );
-                            let _ = app.emit(
-                                "workflow:token_usage",
-                                json!({
-                                    "session_id": session_id,
-                                    "role_id": role,
-                                    "agent_id": agent_id,
-                                    "model": route.model,
-                                    "input_tokens_total": input_total,
-                                    "output_tokens_total": output_total,
-                                    "calls": calls,
-                                }),
-                            );
-                            crate::features::assistant::audit::append(
-                                &ws,
-                                "token",
-                                &role,
-                                json!({
-                                    "agent_id": agent_id,
-                                    "input": usage.input_tokens,
-                                    "output": usage.output_tokens,
-                                }),
-                            );
-                        }
-                        MM::ToolCallStarted {
-                            agent_id,
-                            tool_name,
-                            step,
-                        } => {
-                            // 兜底语义跟 AgentProgress 臂一致(agent_id 而非空串)
-                            let role = agent_roles
-                                .get(&agent_id)
-                                .cloned()
-                                .unwrap_or_else(|| agent_id.clone());
-                            let payload = json!({
-                                "session_id": session_id,
-                                "agent_id": agent_id,
-                                "role_id": role,
-                                "status": format!("🔧 {tool_name} (step {step})"),
-                            });
-                            let _ = app.emit("workflow:agent_progress", payload);
-                        }
-                        MM::Completed { agent_id, summary } => {
-                            turn_shell_tasks.complete_agent(&agent_id);
-                            let role = agent_roles.get(&agent_id).cloned().unwrap_or_default();
-                            crate::features::assistant::audit::append(
-                                &ws,
-                                "agent_done",
-                                &role,
-                                json!({
-                                    "agent_id": agent_id,
-                                    "summary": crate::platform::strings::truncate_utf8(&summary, 600),
-                                }),
-                            );
-                        }
-                        MM::Failed { agent_id, error } => {
-                            turn_shell_tasks.complete_agent(&agent_id);
-                            let role = agent_roles.get(&agent_id).cloned().unwrap_or_default();
-                            crate::features::assistant::audit::append(
-                                &ws,
-                                "agent_failed",
-                                &role,
-                                json!({
-                                    "agent_id": agent_id,
-                                    "error": crate::platform::strings::truncate_utf8(&error, 600),
-                                }),
-                            );
-                        }
-                        MM::Cancelled { agent_id } => {
-                            turn_shell_tasks.complete_agent(&agent_id);
-                        }
-                        MM::Interrupted { agent_id, .. } => {
-                            // CodeWhale 当前不保留可原位恢复的 live task；中断后只能基于
-                            // checkpoint 重新派发，因此它对本轮 Shell scope 同样是终态。
-                            turn_shell_tasks.complete_agent(&agent_id);
-                        }
-                        _ => {}
-                    }
-                }
-                // [pinvou3-fork] 真 SubAgent(Step C)完成 → Step D Gate。executing 期间
-                // 主 session 不发 TurnComplete(它空闲,subagent 在 subagent_manager 跑),
-                // 所以单角色周期的"完成"信号在这里、不在 TurnComplete。
-                Event::AgentComplete {
-                    id,
-                    result,
-                    role: envelope_role,
-                    failed,
-                } => {
-                    // 用户停止整个工作流后，SubAgent 可能在任务 abort 前抢先送达
-                    // 一个完成信封。持久 stop marker 是调度熔断边界：迟到结果可留在
-                    // 项目目录，但绝不能再过 gate、补派下一页或推进下一个角色。
-                    if crate::features::assistant::harness::workflow_is_stopped(
-                        &execution_workspace,
-                    ) {
-                        eprintln!("[harness] ignore late AgentComplete after workflow stop: {id}");
-                        continue;
-                    }
-                    eprintln!(
-                        "[harness] subagent {id} complete ({} chars summary, failed={failed})",
-                        result.len()
-                    );
-                    // role/failed 是新增字段:老的工作流看板只认 agent_id,多带两个字段
-                    // 不影响它;多智能体工位看板要靠这两个判断"谁完成了、成没成功",
-                    // 缺了就只能显示一个没有归属的完成信号。
-                    let payload = json!({
-                        "session_id": session_id,
-                        "agent_id": id,
-                        "role_id": envelope_role
-                            .clone()
-                            .or_else(|| agent_roles.get(&id).cloned()),
-                        "failed": failed,
-                    });
-                    let _ = app.emit("workflow:agent_complete", payload);
-                    let state = store.mode_state(&session_id);
-                    if state.active_skill.is_some() {
-                        // [C2] 完成→节点关联完全用信封 role（SDAN Result.from）。
-                        // harness_phase 已删,无 phase 兜底——正常 workflow subagent
-                        // 派发必带 role(SubAgentAssignment.role)。
-                        let decision_role: Option<String> = envelope_role.clone();
-                        if decision_role.is_none() {
-                            eprintln!("[harness] ⚠️AgentComplete 无 role(非 workflow subagent?) id={id},不推进");
-                        }
-                        // dedup:同一 agent_id 重复完成 → 跳过,防双推进。
-                        // 角色重派(gate 失败/回滚)会得新 agent_id,不会被误跳。
-                        let is_dup = seen_completions
-                            .insert(id.clone(), decision_role.clone().unwrap_or_default())
-                            .is_some();
-                        if is_dup {
-                            eprintln!("[harness] 重复 AgentComplete id={id},跳过");
-                        } else if let Some(role) = decision_role {
-                            // [per_page] role 形如 "slide_writer#p01" = fan-out 成员：
-                            // 记一页 join 计数，未齐则等其余页（不推进）；齐了才对【单一
-                            // 逻辑节点】base_role 验收一次。普通角色 base = role 本身。
-                            let base_for_step: Option<String> = if let Some((base, page_str)) =
-                                role.split_once('#')
-                            {
-                                let page: u32 =
-                                    page_str.trim_start_matches('p').parse().unwrap_or(0);
-                                let ws = execution_workspace.clone();
-                                // [per_page] 先校验该页【真写成】：SSE 超时/放弃的 agent 也
-                                // emit AgentComplete，但只留空壳骨架。空壳不计 done，自动重派
-                                // 该页(带上限)，挡住空壳混入 batch → 避免 gate pagenum_mismatch
-                                // 误判回滚的死循环。
-                                let outs = crate::features::assistant::harness::batch_outputs_for(
-                                    &session_id,
-                                    base,
-                                    page,
-                                );
-                                let real = crate::features::assistant::harness::page_output_is_real(
-                                    &ws, base, page, &outs,
-                                );
-                                let mut count_done = real;
-                                if real {
-                                    eprintln!("[harness] per_page {role} 真写成 ✓");
-                                    crate::features::assistant::harness::fanout_mark(
-                                        &session_id,
-                                        base,
-                                        page,
-                                        "done",
-                                    );
-                                } else {
-                                    let n = crate::features::assistant::harness::page_retry_inc(
-                                        &session_id,
-                                        base,
-                                        page,
-                                    );
-                                    let maxr =
-                                        crate::features::assistant::harness::max_page_retry();
-                                    if n <= maxr {
-                                        // 空壳 → 重派该页(占用刚释放的在飞名额，不补排队页)。
-                                        let ws_r = ws.clone();
-                                        let base_r = base.to_string();
-                                        let respawn = tokio::task::spawn_blocking(move || {
-                                            crate::features::assistant::harness::respawn_page(
-                                                &ws_r, &base_r, page,
-                                            )
-                                        })
-                                        .await
-                                        .unwrap_or(None);
-                                        if let Some(t) = respawn {
-                                            let rr = t.agent_role.clone();
-                                            let op = Op::SpawnSubAgent {
-                                                prompt: t.prompt,
-                                                role_id: t.agent_role,
-                                                allowed_tools: t.allowed_tools,
-                                                write_files: t.write_files,
-                                                max_steps: t.max_steps,
-                                                output_schema: t.output_schema,
-                                                structured_output_root: Some(t.project_dir),
-                                                expects_file_output: t.expects_file_output,
-                                            };
-                                            if let Err(e) = approve_handle.send(op).await {
-                                                eprintln!("[harness] per_page {role} 空壳→重派 {rr} 失败: {e:?}");
-                                            } else {
-                                                eprintln!("[harness] per_page {role} 空壳→重派(第{n}/{maxr}次) {rr}");
-                                                crate::features::assistant::harness::fanout_mark(
-                                                    &session_id,
-                                                    base,
-                                                    page,
-                                                    "retrying",
-                                                );
-                                            }
-                                        } else {
-                                            eprintln!("[harness] per_page {role} 空壳但 respawn 取不到任务 → 兜底计 done");
-                                            count_done = true;
-                                            crate::features::assistant::harness::fanout_mark(
-                                                &session_id,
-                                                base,
-                                                page,
-                                                "done",
-                                            );
-                                        }
-                                    } else {
-                                        eprintln!("[harness] per_page {role} 空壳且重试耗尽({maxr}) → 兜底计 done(留给 gate/人工)");
-                                        count_done = true;
-                                        crate::features::assistant::harness::fanout_mark(
-                                            &session_id,
-                                            base,
-                                            page,
-                                            "done",
-                                        );
-                                    }
-                                }
-                                // 每页完成/重试后把最新 fan-out 状态推给前端工作流界面。
-                                emit_fanout(&app, &session_id, base);
-                                if count_done {
-                                    let ws_d = ws.clone();
-                                    let base_d = base.to_string();
-                                    let complete = tokio::task::spawn_blocking(move || {
-                                        crate::features::assistant::harness::record_page_done(
-                                            &ws_d, &base_d, page,
-                                        )
-                                    })
-                                    .await
-                                    .unwrap_or(true);
-                                    eprintln!(
-                                        "[harness] per_page {role} done; batch_complete={complete}"
-                                    );
-                                    if complete {
-                                        crate::features::assistant::harness::batch_clear(
-                                            &session_id,
-                                            base,
-                                        );
-                                        Some(base.to_string())
-                                    } else {
-                                        // 未齐 → 补派下一排队页，维持在飞并发=K；不推进。
-                                        if let Some(t) =
-                                            crate::features::assistant::harness::batch_pop_next(
-                                                &session_id,
-                                                base,
-                                            )
-                                        {
-                                            let next_role = t.agent_role.clone();
-                                            let op = Op::SpawnSubAgent {
-                                                prompt: t.prompt,
-                                                role_id: t.agent_role,
-                                                allowed_tools: t.allowed_tools,
-                                                write_files: t.write_files,
-                                                max_steps: t.max_steps,
-                                                output_schema: t.output_schema,
-                                                structured_output_root: Some(t.project_dir),
-                                                expects_file_output: t.expects_file_output,
-                                            };
-                                            if let Err(e) = approve_handle.send(op).await {
-                                                eprintln!("[harness] per_page 补派 {next_role} 失败: {e:?}");
-                                            } else {
-                                                eprintln!(
-                                                    "[harness] per_page 补派下一页 {next_role}"
-                                                );
-                                            }
-                                        }
-                                        None
-                                    }
-                                } else {
-                                    // 重派路径：等该页重试结果，不推进、不补排队页。
-                                    None
-                                }
-                            } else {
-                                Some(role.clone())
-                            };
-
-                            if let Some(base_role) = base_for_step {
-                                let ws = execution_workspace.clone();
-                                // [2026-06-06] SubAgent 执行失败(failed=true,单角色)绝不走 gate：
-                                // gate 只验产物存在+非空，会拿【上一轮陈旧产物】把失败洗成 PASS
-                                // (实锤:web_search 不可用→PM 0步即死→旧 brief 过关)。改走
-                                // agent_failed:--fail 计次→重派带失败原因，耗尽→Blocked。
-                                // per_page 成员(role 带 #)不走这里:空壳检测已按页处理。
-                                let failed_single = failed && !role.contains('#');
-                                let err_text = result.clone();
-                                let action = tokio::task::spawn_blocking(move || {
-                                    if failed_single {
-                                        crate::features::assistant::harness::agent_failed(
-                                            &ws, &base_role, &err_text,
-                                        )
-                                    } else {
-                                        crate::features::assistant::harness::step_after_role(
-                                            &ws, &base_role,
-                                        )
-                                    }
-                                })
-                                .await
-                                .unwrap_or_else(|_| {
-                                    crate::features::assistant::harness::HarnessAction::Error(
-                                        "spawn_blocking panicked".into(),
-                                    )
-                                });
-                                let handled = apply_harness_action(
-                                    action,
-                                    &app,
-                                    &execution_workspace,
-                                    &approve_handle,
-                                    &session_id,
-                                )
-                                .await;
-                                if handled {
-                                    let ws = execution_workspace.clone();
-                                    let app_clone = app.clone();
-                                    let sid_clone = session_id.clone();
-                                    tokio::task::spawn_blocking(move || {
-                                        if let Some(mut st) =
-                                            crate::features::assistant::harness::read_full_agent_state(&ws)
-                                        {
-                                            if let Some(obj) = st.as_object_mut() {
-                                                obj.insert("session_id".into(), json!(sid_clone));
-                                            }
-                                            let _ = app_clone.emit("workflow:full_state", st);
-                                        }
-                                    });
-                                }
-                            }
-                        }
-                    }
-                }
-                Event::TurnComplete {
-                    usage,
-                    status,
-                    error,
-                    // v0.8.49 上游新增 tool_catalog / base_url(调试/审计用),pinvou3 不消费
-                    ..
-                } => {
-                    let shell_turn_id = current_turn_id.clone();
-                    let shell_interrupted = status == TurnOutcomeStatus::Interrupted;
-                    let mut shell_cleanup_failed = false;
-                    let mut terminal_status = status;
-                    let mut terminal_error = error;
-                    if let Some(base_total_tokens) = scheduled_base_total_tokens {
-                        scheduled_engine_total_tokens = scheduled_engine_total_tokens
-                            .saturating_add(u64::from(usage.input_tokens))
-                            .saturating_add(u64::from(usage.output_tokens));
-                        let store_for_save = store.clone();
-                        let session_for_save = session_id.clone();
-                        match tokio::task::spawn_blocking(move || {
-                            store_for_save.persist_scheduled_token_total(
-                                &session_for_save,
-                                base_total_tokens,
-                                scheduled_engine_total_tokens,
-                            )
-                        })
-                        .await
-                        {
-                            Ok(Ok(_)) => {}
-                            Ok(Err(save_error)) => {
-                                scheduled_persistence_error = Some(format!("{save_error:#}"));
-                            }
-                            Err(join_error) => {
-                                scheduled_persistence_error = Some(format!(
-                                    "scheduled token persistence task failed: {join_error}"
-                                ));
-                            }
-                        }
-                        if let Some(save_error) = scheduled_persistence_error.take() {
-                            terminal_status = TurnOutcomeStatus::Failed;
-                            terminal_error = Some(match terminal_error {
-                                Some(engine_error) => format!(
-                                    "{engine_error}; scheduled conversation persistence failed: {save_error}"
-                                ),
-                                None => format!(
-                                    "Scheduled conversation persistence failed: {save_error}"
-                                ),
-                            });
-                        }
-                    } else {
-                        let terminal_save = if active_transcript_seen {
-                            latest_chat_engine_state.clone().map(|state| {
-                                let store_for_save = store.clone();
-                                let session_for_save = session_id.clone();
-                                tokio::task::spawn_blocking(move || {
-                                    store_for_save
-                                        .persist_chat_engine_state(&session_for_save, state)
-                                })
-                            })
-                        } else {
-                            turn_lifecycle.active_transcript_fallback().map(|fallback| {
-                                let store_for_save = store.clone();
-                                let session_for_save = session_id.clone();
-                                tokio::task::spawn_blocking(move || {
-                                    store_for_save.persist_admitted_chat_display(
-                                        &session_for_save,
-                                        &fallback.baseline_revision,
-                                        fallback.display_message,
-                                        fallback.operation == TranscriptOperation::EditLast,
-                                    )
-                                })
-                            })
-                        };
-                        if let Some(save_task) = terminal_save {
-                            match save_task.await {
-                                Ok(Ok(saved)) => {
-                                    chat_persistence_error = None;
-                                    if let Ok(revision) = transcript_revision(&saved.messages) {
-                                        let payload = json!({
-                                            "session_id": session_id,
-                                            "transcript_revision": revision,
-                                        });
-                                        let _ =
-                                            app.emit("chat:transcript_committed", payload.clone());
-                                        crate::features::remote_control::forward_app_event(
-                                            &app,
-                                            "chat:transcript_committed",
-                                            payload,
-                                        );
-                                    }
-                                }
-                                Ok(Err(save_error)) => {
-                                    chat_persistence_error = Some(format!("{save_error:#}"));
-                                }
-                                Err(join_error) => {
-                                    chat_persistence_error = Some(format!(
-                                        "chat transcript persistence task failed: {join_error}"
-                                    ));
-                                }
-                            }
-                        }
-                        if let Some(save_error) = chat_persistence_error.take() {
-                            terminal_status = TurnOutcomeStatus::Failed;
-                            terminal_error = Some(match terminal_error {
-                                Some(engine_error) => format!(
-                                    "{engine_error}; chat conversation persistence failed: {save_error}"
-                                ),
-                                None => format!(
-                                    "Chat conversation persistence failed: {save_error}"
-                                ),
-                            });
-                        }
-                    }
-                    // Keep the shell scope cancellable throughout persistence.
-                    // Final cleanup runs before terminal admission is claimed;
-                    // Engine reclaim can therefore still win an await race and
-                    // close the same scope without leaving the UI permanently
-                    // busy. The cleanup itself executes on blocking workers.
-                    if let Some(turn_id) = shell_turn_id.as_deref() {
-                        match turn_shell_tasks
-                            .finalize_turn(turn_id, shell_interrupted)
-                            .await
-                        {
-                            Ok(report) => {
-                                if !report.killed.is_empty() {
-                                    log::info!(
-                                        "[pinvou3][chat] finalized {} interrupted shell task(s) sid={} turn={}",
-                                        report.killed.len(),
-                                        session_id,
-                                        turn_id
-                                    );
-                                }
-                                if let Some(cleanup_error) = report.failure_summary() {
-                                    shell_cleanup_failed = true;
-                                    log::error!(
-                                        "[pinvou3][chat] shell cleanup remained incomplete sid={} turn={}: {}",
-                                        session_id,
-                                        turn_id,
-                                        cleanup_error
-                                    );
-                                }
-                            }
-                            Err(error) => {
-                                log::error!(
-                                    "[pinvou3][chat] failed to finalize shell task scope sid={} turn={}: {error:#}",
-                                    session_id,
-                                    turn_id
-                                );
-                                shell_cleanup_failed = true;
-                            }
-                        }
-                    }
-                    // Persistence above can change the authoritative outcome and
-                    // may await. Claim only after that result is known, but before
-                    // every completion-only side effect below. If reclaim won
-                    // during the await, this late TurnComplete is discarded.
-                    let Some(_) = turn_lifecycle.claim_terminal_with_admission(&app, &session_id)
-                    else {
-                        continue;
-                    };
-                    // 单独发 usage 给前端 token 进度条
-                    let payload = json!({
-                        "session_id": session_id,
-                        "input_tokens": usage.input_tokens,
-                        "output_tokens": usage.output_tokens,
-                        "context_window": bridge.usage_context_window(),
-                    });
-                    let _ = app.emit("chat:usage", payload.clone());
-                    crate::features::remote_control::forward_app_event(&app, "chat:usage", payload);
-                    // app 侧自测:用精确 usage 收尾本轮。部分远端模型会流式返回文本,
-                    // 但最终 usage.output_tokens 为 0,因此不能用 output_tokens 门控收尾。
-                    if let Some(m) = &self_metrics {
-                        m.on_turn_complete(
-                            &session_id,
-                            usage.input_tokens,
-                            usage.output_tokens,
-                            usage.prompt_cache_hit_tokens,
-                            usage.prompt_cache_miss_tokens,
-                        );
-                    }
-                    // turn end:取出 tracker 快照,然后重置(下个 turn 重新累积)。
-                    // 用独立 block 把 parking_lot guard 的生命周期限死在这里:下方
-                    // H1 harness 段有 .await(spawn_blocking),guard 是 !Send,若跨
-                    // await 会让整个 forwarder future 变 !Send(tauri::spawn 要求 Send)。
-                    let (plan_used, plan_snapshot, todos_snapshot) = {
-                        let mut tracker = plan_tracker.lock();
-                        let plan_used = tracker.plan_tool_used;
-                        let plan_snapshot = tracker.last_plan_snapshot.take();
-                        let todos_snapshot = tracker.last_todos_snapshot.take();
-                        *tracker = TurnPlanTracker::default();
-                        (plan_used, plan_snapshot, todos_snapshot)
-                    };
-
-                    // 先完成权威时间线，再发 chat:done。前端终态事件会据此重新
-                    // 读取整份时间线，补齐后台运行/页面恢复期间漏掉的本地事件。
-                    let status_text = format!("{terminal_status:?}");
-                    crate::features::assistant::timing::finish_turn_with_usage(
-                        &session_id,
-                        &status_text,
-                        terminal_error.as_deref(),
-                        Some(crate::features::assistant::timing::TurnUsage {
-                            input_tokens: u64::from(usage.input_tokens),
-                            output_tokens: u64::from(usage.output_tokens),
-                            cache_hit_tokens: u64::from(usage.prompt_cache_hit_tokens.unwrap_or(0)),
-                            cache_miss_tokens: u64::from(
-                                usage.prompt_cache_miss_tokens.unwrap_or(0),
-                            ),
-                            cache_write_tokens: u64::from(
-                                usage.prompt_cache_write_tokens.unwrap_or(0),
-                            ),
-                            reasoning_tokens: u64::from(usage.reasoning_tokens.unwrap_or(0)),
-                        }),
-                    );
-
-                    {
-                        // 多引擎:mode 判据基于本 forwarder 的 session_id,不读全局 active。
-                        let active_id = session_id.clone();
-                        let state = store.mode_state(&active_id);
-
-                        // ── plan_ready 触发(Plan + 调过 plan 类工具) ──
-                        // 底座式:Plan 模式调过 update_plan = 出方案 → 弹决策卡。不再设 phase
-                        // (已砍 PlanPhase),mode 留 Plan 直到用户 accept/discard 切 Yolo。
-                        // 修订(还在 Plan 时再调 update_plan)天然幂等:再弹一张新卡。
-                        if plan_used && state.mode == SerializableMode::Plan {
-                            if let Some(plan_id) = current_turn_id.clone() {
-                                if let Some(mode_state) =
-                                    store.register_pending_plan(&active_id, plan_id.clone())
-                                {
-                                    let payload = json!({
-                                        "session_id": active_id.clone(),
-                                        "plan_id": plan_id,
-                                        "plan_snapshot": plan_snapshot.clone(),
-                                        "todos_snapshot": todos_snapshot.clone(),
-                                        "mode_state": mode_state,
-                                    });
-                                    let _ = app.emit("chat:plan_ready", payload.clone());
-                                    crate::features::remote_control::forward_app_event(
-                                        &app,
-                                        "chat:plan_ready",
-                                        payload,
-                                    );
-                                }
-                            }
-                        }
-
-                        // From here the harness path may await. Publish the claimed
-                        // terminal first so a concurrent Engine eviction cannot
-                        // abort this forwarder and leave the frontend permanently
-                        // busy. Reclaim now observes the lifecycle as terminal and
-                        // cannot publish a conflicting Interrupted outcome.
-                        emit_chat_terminal(
-                            &app,
-                            &session_id,
-                            terminal_status,
-                            terminal_error.clone(),
-                            shell_cleanup_failed,
-                        );
-                        turn_lifecycle.finish_terminal_emission();
-
-                        // ── H1: Harness Loop (skill session 图执行器) ──
-                        // 本 session 绑定了 skill 且项目目录有 workflow_progress.json →
-                        // harness 图执行器始终使用 engine 的实际执行工作区：普通会话是
-                        // session 私有目录，定时会话是所属 automation 的专属工作间。
-                        let harness_workspace = execution_workspace.clone();
-                        let harness_handled = if state.active_skill.is_some() {
-                            // [C2] harness_phase 已删。工作流由 kick(命令)+ AgentComplete
-                            // 驱动;主 session 在工作流期间空闲,此处 TurnComplete 一般不参与
-                            // 推进。仍兜底走 step_fresh:由 scheduler 据 State 决策——有角色在
-                            // 跑则返回 role_running→NotApplicable(防重复派发),否则派下一个。
-                            let ws = harness_workspace.clone();
-                            let action = tokio::task::spawn_blocking(move || {
-                                crate::features::assistant::harness::step_fresh(&ws)
-                            })
-                            .await
-                            .unwrap_or(
-                                crate::features::assistant::harness::HarnessAction::Error(
-                                    "spawn_blocking panicked".into(),
-                                ),
-                            );
-
-                            apply_harness_action(
-                                action,
-                                &app,
-                                &execution_workspace,
-                                &approve_handle,
-                                &active_id,
-                            )
-                            .await
-                        } else {
-                            false
-                        };
-
-                        // ── H1b: harness 推进了 → 推送全量 agent 状态快照给前端 ──
-                        if harness_handled {
-                            let ws = harness_workspace.clone();
-                            let app_clone = app.clone();
-                            let sid_clone = active_id.clone();
-                            tokio::task::spawn_blocking(move || {
-                                if let Some(mut state) =
-                                    crate::features::assistant::harness::read_full_agent_state(&ws)
-                                {
-                                    if let Some(obj) = state.as_object_mut() {
-                                        obj.insert("session_id".into(), json!(sid_clone));
-                                    }
-                                    let _ = app_clone.emit("workflow:full_state", state);
-                                }
-                            });
-                        }
-
-                        // ── [回归底座式] M2 自驱 + M3 文本兜底已彻底砍掉 ──
-                        // M2:执行不自动续跑,回底座由用户驱动。M3(Plan 写了方案没调
-                        // update_plan 的救援)放弃不做:底座 update_plan→plan_ready→方案卡
-                        // 这条链已可靠,漏的少数"光说不出卡"由 composer chip 手切 + plan_stuck
-                        // 卡兜底,不值得用噪音判据再造一层。
-                    }
-                    if crate::features::memory::memory_enabled() {
-                        if let Some(capture) =
-                            crate::features::memory::take_turn_capture(&session_id)
-                        {
-                            let app_clone = app.clone();
-                            let bridge_clone = bridge.clone();
-                            let sid_clone = session_id.clone();
-                            tauri::async_runtime::spawn(async move {
-                                match crate::features::memory::review_turn_candidates_with_llm(
-                                    &bridge_clone,
-                                    &capture,
-                                    &sid_clone,
-                                )
-                                .await
-                                {
-                                    Ok(outcome) => {
-                                        let mut events = outcome.events;
-                                        events.extend(outcome.pending.into_iter().filter_map(
-                                            |item| {
-                                                (item.status == "pending_confirm").then(|| {
-                                                    crate::features::memory::MemoryWriteEvent {
-                                                        kind: item.kind,
-                                                        action: "pending".to_string(),
-                                                        id: item.id,
-                                                        text: item.content,
-                                                    }
-                                                })
-                                            },
-                                        ));
-                                        if !events.is_empty() {
-                                            let _ = app_clone.emit(
-                                                "chat:memory_write",
-                                                json!({
-                                                    "session_id": sid_clone,
-                                                    "events": events,
-                                                }),
-                                            );
-                                            match crate::features::memory::runtime_snapshot(&sid_clone) {
-                                            Ok(snapshot) => {
-                                                let _ = app_clone.emit(
-                                                    "chat:memory",
-                                                    json!({
-                                                        "session_id": sid_clone,
-                                                        "items": snapshot.items,
-                                                        "runtime_path": snapshot.runtime_path,
-                                                    }),
-                                                );
-                                            }
-                                            Err(err) => eprintln!(
-                                                "[pinvou3-app] refresh memory runtime after review failed for session {sid_clone}: {err}"
-                                            ),
-                                        }
-                                        }
-                                    }
-                                    Err(err) => {
-                                        eprintln!(
-                                        "[pinvou3-app] memory llm review failed for session {sid_clone}: {err:#}"
-                                    );
-                                    }
-                                }
-                            });
-                        }
-                    }
-                    maybe_notify_task_completed(
-                        &app,
-                        &store,
-                        &session_id,
-                        current_turn_id.take(),
-                        terminal_status,
-                        terminal_error.as_deref(),
-                    );
-                    if let Some(signal) = turn_tracker.on_terminal(terminal_status, terminal_error)
-                    {
-                        let _ = turn_events.send(signal);
-                    }
-                }
-                Event::CompactionStarted {
-                    id, message, auto, ..
-                } => {
-                    let payload = json!({ "session_id": session_id, "phase": "start", "id": id, "auto": auto, "message": message });
-                    let _ = app.emit("chat:compaction", payload.clone());
-                    crate::features::remote_control::forward_app_event(
-                        &app,
-                        "chat:compaction",
-                        payload,
-                    );
-                }
-                Event::CompactionCompleted {
-                    id,
-                    message,
-                    auto,
-                    messages_before,
-                    messages_after,
-                    ..
-                } => {
-                    let payload = json!({
-                        "session_id": session_id,
-                        "phase": "done",
-                        "id": id,
-                        "auto": auto,
-                        "message": message,
-                        "messages_before": messages_before,
-                        "messages_after": messages_after,
-                    });
-                    let _ = app.emit("chat:compaction", payload.clone());
-                    crate::features::remote_control::forward_app_event(
-                        &app,
-                        "chat:compaction",
-                        payload,
-                    );
-                }
-                Event::CompactionFailed { message, auto, .. } => {
-                    let payload = json!({ "session_id": session_id, "phase": "fail", "auto": auto, "message": message });
-                    let _ = app.emit("chat:compaction", payload.clone());
-                    crate::features::remote_control::forward_app_event(
-                        &app,
-                        "chat:compaction",
-                        payload,
-                    );
-                }
-                Event::Error { envelope, .. } => {
-                    // 可恢复错误(如 SSE idle timeout、瞬态工具失败)turn 不会结束——
-                    // 引擎会 retry / 继续跑后续步骤。绝不能发 chat:done,否则前端
-                    // setBusy(false) 把"思考中"指示器掐掉,而引擎还在干活,看着像卡死
-                    // (且会误触发 flush/closeBubble/plan_phase 收尾)。只飘个 advisory。
-                    // 仅 recoverable==false(致命)才是真结束 → chat:done。
-                    if envelope.recoverable {
-                        let payload =
-                            json!({ "session_id": session_id, "error": envelope.message });
-                        let _ = app.emit("chat:transient_error", payload.clone());
-                        crate::features::remote_control::forward_app_event(
-                            &app,
-                            "chat:transient_error",
-                            payload,
-                        );
-                    } else {
-                        // 底座在致命 Error 后仍会发权威 TurnComplete(Failed)。这里只缓存
-                        // 文本；完成、持久化和 chat:done 全部由 TurnComplete 单点处理。
-                        // [C2] harness_phase 已删。工作流绑定(active_skill)的 session 遇
-                        // 致命错误(SubAgent 派发失败 / 内部 fatal)→ 可能收不到 AgentComplete
-                        // = 死锁。兜底:emit blocked 通知前端,让用户看到中断可重开(宁可多
-                        // 通知一次,不可无声卡死等永不到来的事件)。
-                        let was_active = store.mode_state(&session_id).active_skill.is_some();
-                        if was_active {
-                            let _ = app.emit(
-                                "workflow:blocked",
-                                json!({
-                                    "session_id": session_id,
-                                    "message": "工作流执行中断（致命错误，可能是 SubAgent 派发失败或内部错误），请检查后重新开始。",
-                                }),
-                            );
-                        }
-                        turn_tracker.on_fatal_error(envelope.message);
-                    }
-                }
-                _ => {}
-            }
-        }
-        let stopped_error = "Engine event stream stopped before a terminal event".to_string();
-        let mut shell_cleanup_failed = false;
-        {
-            // The stream can stop before `TurnStarted` bound the provisional
-            // submission scope; close the active scope too, otherwise the next
-            // `prepare_turn` keeps bailing on "scope already active".
-            let finalize = match current_turn_id.as_deref() {
-                Some(turn_id) => turn_shell_tasks.finalize_turn(turn_id, true).await,
-                None => turn_shell_tasks.finalize_active_scope(true).await,
-            };
-            match finalize {
-                Ok(report) => {
-                    if let Some(error) = report.failure_summary() {
-                        shell_cleanup_failed = true;
-                        log::error!(
-                            "[pinvou3][chat] shell cleanup remained incomplete after event stream stop sid={} turn={:?}: {}",
-                            session_id,
-                            current_turn_id,
-                            error
-                        );
-                    }
-                }
-                Err(error) => {
-                    shell_cleanup_failed = true;
-                    log::error!(
-                        "[pinvou3][chat] failed to close shell task scope after event stream stop sid={} turn={:?}: {error:#}",
-                        session_id,
-                        current_turn_id
-                    );
-                }
-            }
-        }
-        match finish_reclaimed_lifecycle_turn(
-            &turn_lifecycle,
-            &app,
-            &store,
-            &session_id,
-            TurnOutcomeStatus::Failed,
-            Some(stopped_error.clone()),
-            shell_cleanup_failed,
-        )
-        .await
-        {
-            Some(EmittedTerminal {
-                turn_id: Some(turn_id),
-            }) => {
-                let _ = turn_events.send(EngineTurnSignal::Terminal {
-                    turn_id,
-                    status: TurnOutcomeStatus::Failed,
-                    error: Some(stopped_error),
-                });
-            }
-            Some(EmittedTerminal { turn_id: None }) | None => {
-                let _ = turn_events.send(EngineTurnSignal::ForwarderStopped {
-                    error: stopped_error,
-                });
-            }
-        }
-        eprintln!(
-            "[pinvou3-app] event forwarder stopped for session {session_id} (engine shut down?)"
-        );
-    })
-}
+#[path = "forwarder.rs"]
+mod forwarder;
+pub(crate) use forwarder::spawn_event_forwarder;
 
 /// 让 main.rs 编译时知道这个模块（供 docs/CI 用）。
 pub fn _force_link() -> Arc<()> {
@@ -3354,7 +1778,6 @@ mod turn_lifecycle_tests {
     use crate::core::mode_state::SessionModeState;
     use deepseek_tui::models::{ContentBlock, Message};
     use std::cell::Cell;
-    use std::sync::atomic::{AtomicBool, Ordering};
     use std::sync::Arc;
 
     fn message(role: &str, text: &str) -> Message {
@@ -3488,19 +1911,18 @@ mod turn_lifecycle_tests {
 
         // reserve 后 active=true 但 submitted=false → arm 不得置位。
         let reservation = lifecycle.reserve().expect("reserve");
-        let epoch = lifecycle.current_turn_generation().expect("active epoch");
-        lifecycle.arm_pending_cancel_and_cancel(epoch, || {});
+        lifecycle.arm_pending_cancel();
         assert!(
-            lifecycle.take_pending_cancel(epoch).is_none(),
+            !lifecycle.take_pending_cancel(),
             "must not arm pending_cancel for an unsubmitted reservation"
         );
 
         // 走真实 send 路径：handle.send 成功后 mark_submitted → submitted=true。
         lifecycle.mark_reservation_submitted(reservation.reservation_id);
         // 已 submitted + active + turn_id=None（TurnStarted 未抵达）→ arm 置位。
-        lifecycle.arm_pending_cancel_and_cancel(epoch, || {});
+        lifecycle.arm_pending_cancel();
         assert!(
-            lifecycle.take_pending_cancel(epoch).is_some(),
+            lifecycle.take_pending_cancel(),
             "pending_cancel must be armed after submission, before TurnStarted"
         );
         // 消费 reservation 避免 Drop 副作用。
@@ -3545,303 +1967,29 @@ mod turn_lifecycle_tests {
         // 3. 已 started 的 turn 不 arm（cancel_current 直接命中活跃 token）。
         let lifecycle = Arc::new(TurnLifecycle::default());
 
-        // --- 场景 A：submit 后、TurnStarted 前 arm → take 返回 Some ---
+        // --- 场景 A：submit 后、TurnStarted 前 arm → take 返回 true ---
         lifecycle.on_submitted();
-        let epoch_a = lifecycle.current_turn_generation().expect("active epoch");
-        lifecycle.arm_pending_cancel_and_cancel(epoch_a, || {});
+        lifecycle.arm_pending_cancel();
         assert!(
-            lifecycle.take_pending_cancel(epoch_a).is_some(),
+            lifecycle.take_pending_cancel(),
             "pending_cancel must be armed before TurnStarted"
         );
-        // take 已消费，再次取返回 None。
+        // take 已消费，再次取返回 false。
         assert!(
-            lifecycle.take_pending_cancel(epoch_a).is_none(),
+            !lifecycle.take_pending_cancel(),
             "pending_cancel must be consumed exactly once"
         );
 
         // --- 场景 B：TurnStarted 后 arm → 不置标记 ---
         lifecycle.on_started("turn-1".to_string());
-        let epoch_b = lifecycle.current_turn_generation().expect("active epoch");
-        lifecycle.arm_pending_cancel_and_cancel(epoch_b, || {});
+        lifecycle.arm_pending_cancel();
         assert!(
-            lifecycle.take_pending_cancel(epoch_b).is_none(),
+            !lifecycle.take_pending_cancel(),
             "must not arm pending_cancel after TurnStarted"
         );
 
         // 清理：结束当前 turn。
         assert!(lifecycle.finish_once(|| {}).is_some());
-    }
-
-    #[test]
-    fn arm_pending_cancel_uses_target_snapshot_not_reread_current() {
-        // reviewer 点 3 的确定性回归：cancel 在 generation 复查通过后 arm
-        // pending 时，必须使用发起时快照 target，而不是此刻重读的当前 epoch。
-        // 若重读 current：复查通过后、arm 前另一 worker 结束旧轮并启动新轮，
-        // 读到的是新轮 epoch，arm_pending_cancel(新epoch) 会把 pending 绑定到
-        // 新轮（新轮恰好处于 submitted 未 started 时），后面的 cancel_current
-        // 命中新 token → 原始跨轮误取消窗口仍未关闭。
-        //
-        // 这里直接编排「复查通过后轮次切换」：turn1 处于 submitted 未 started
-        // （arm 前置满足），快照 target=epoch1；切换 turn2 并提交后，用 target
-        // arm 必须被拒绝（turn_epoch 已 != target），pending 不落到新轮。
-        let lifecycle = Arc::new(TurnLifecycle::default());
-
-        // turn1：on_submitted 激活（active+submitted+epoch=1，turn_id=None）。
-        assert!(lifecycle.on_submitted());
-        let target = lifecycle
-            .current_turn_generation()
-            .expect("turn1 active epoch");
-
-        // 复查通过后、arm 前轮次切换：turn1 终态 → turn2 reserve + 提交。
-        assert!(lifecycle.finish_once(|| {}).is_some());
-        let reservation2 = lifecycle.reserve().expect("turn2 reserve");
-        lifecycle.mark_reservation_submitted(reservation2.reservation_id);
-        let epoch2 = lifecycle
-            .current_turn_generation()
-            .expect("turn2 active epoch");
-        assert_eq!(epoch2, target + 1, "turn2 must bump the epoch past target");
-
-        // 用发起时快照 target arm：turn_epoch 已是 epoch2 != target → 拒绝并
-        // 返回 false（reviewer 点 6：调用方据返回值跳过 cancel_current，否则
-        // 取消会命中新轮已 reset_cancel_token 的活跃 token）。
-        assert!(
-            !lifecycle.arm_pending_cancel_and_cancel(target, || {}),
-            "arm with the stale target snapshot must report epoch mismatch"
-        );
-        assert!(
-            lifecycle.take_pending_cancel(epoch2).is_none(),
-            "arm with the stale target snapshot must not bind pending to the new turn"
-        );
-
-        // 防 Drop 副作用。
-        reservation2.mark_submitted();
-    }
-
-    #[test]
-    fn arm_pending_cancel_reports_epoch_mismatch_and_arms_current() {
-        // reviewer 点 6 的原子边界：arm 在 state 锁内校验 turn_epoch == epoch。
-        // 轮次已切换 → 返回 false 且不置 pending（调用方必须跳过取消副作用）；
-        // epoch 匹配 → 返回 true 并正常 arm（submitted 未 started 时置 pending）。
-        let lifecycle = Arc::new(TurnLifecycle::default());
-
-        // turn1：on_submitted 激活（active+submitted+epoch=1，turn_id=None）。
-        assert!(lifecycle.on_submitted());
-        let target = lifecycle
-            .current_turn_generation()
-            .expect("turn1 active epoch");
-
-        // 轮次切换：turn1 终态 → turn2 reserve + 提交（epoch=2）。
-        assert!(lifecycle.finish_once(|| {}).is_some());
-        let reservation2 = lifecycle.reserve().expect("turn2 reserve");
-        lifecycle.mark_reservation_submitted(reservation2.reservation_id);
-        let epoch2 = lifecycle
-            .current_turn_generation()
-            .expect("turn2 active epoch");
-        assert_eq!(epoch2, target + 1, "turn2 must bump the epoch past target");
-
-        // stale target：拒绝 + pending 不落到新轮。
-        assert!(!lifecycle.arm_pending_cancel_and_cancel(target, || {}));
-        assert!(
-            lifecycle.take_pending_cancel(epoch2).is_none(),
-            "rejected arm must not set pending on the new turn"
-        );
-
-        // 当前 epoch：接受 + pending 置位（turn2 仍 submitted 未 started）。
-        assert!(lifecycle.arm_pending_cancel_and_cancel(epoch2, || {}));
-        assert!(
-            lifecycle.take_pending_cancel(epoch2).is_some(),
-            "accepted arm must set pending for the current turn"
-        );
-
-        // 防 Drop 副作用。
-        reservation2.mark_submitted();
-    }
-
-    #[test]
-    fn arm_pending_cancel_and_cancel_holds_state_lock_through_cancel() {
-        // reviewer 点 8 的确定性回归：epoch 校验/arm 与取消副作用必须在同一
-        // lifecycle state 锁临界区内原子完成。若 arm 后释放锁、再单独执行
-        // cancel_current，两条同步调用之间没有 await 也不构成原子性保证——
-        // 另一 worker 可在该窗口内收口旧轮、reserve/发送新轮并
-        // reset_cancel_token，恢复后旧 cancel 命中新轮活跃 token。
-        //
-        // 编排：cancel 闭包（持锁）进入后阻塞，期间另一线程尝试 reserve——
-        // 必须被 state 锁挡住；放行 cancel 后 reserve 才得以完成。若实现把
-        // cancel 放在锁外，reserve 会在 cancel 完成前成功，断言失败。
-        let lifecycle = Arc::new(TurnLifecycle::default());
-        // turn1：on_submitted 激活（active+submitted+epoch=1），cancel 目标轮。
-        assert!(lifecycle.on_submitted());
-        let target = lifecycle.current_turn_generation().expect("turn1 epoch");
-
-        let (entered_tx, entered_rx) = std::sync::mpsc::channel::<()>();
-        let (release_tx, release_rx) = std::sync::mpsc::channel::<()>();
-
-        // cancel 线程：arm_pending_cancel_and_cancel 持锁执行 cancel 闭包，
-        // 闭包通知「已进入临界区」后阻塞，模拟取消副作用执行中。
-        let lc_for_cancel = lifecycle.clone();
-        let cancel_thread = std::thread::spawn(move || {
-            let ok = lc_for_cancel.arm_pending_cancel_and_cancel(target, || {
-                entered_tx.send(()).expect("entered");
-                release_rx.recv().expect("release");
-            });
-            assert!(ok, "epoch must match when the cancel side effect runs");
-        });
-
-        // 等 cancel 进入临界区（已持 state 锁）。
-        entered_rx.recv().expect("cancel entered");
-
-        // 另一 worker 尝试 reserve 新轮：应被 state 锁阻塞，直到 cancel 完成。
-        let (reserve_done_tx, reserve_done_rx) = std::sync::mpsc::channel::<()>();
-        let lc_for_reserve = lifecycle.clone();
-        let reserve_thread = std::thread::spawn(move || {
-            let _ = lc_for_reserve.reserve();
-            reserve_done_tx.send(()).expect("reserve done");
-        });
-
-        // 确定性断言：cancel 持锁期间 reserve 不能完成。
-        assert!(
-            reserve_done_rx
-                .recv_timeout(std::time::Duration::from_millis(200))
-                .is_err(),
-            "reserve must be blocked while the cancel side effect holds the state lock"
-        );
-
-        // 放行 cancel → reserve 随后完成。
-        release_tx.send(()).expect("release cancel");
-        cancel_thread.join().expect("cancel thread joins");
-        reserve_thread.join().expect("reserve thread joins");
-        assert!(
-            reserve_done_rx
-                .recv_timeout(std::time::Duration::from_secs(1))
-                .is_ok(),
-            "reserve must complete after the cancel side effect releases the state lock"
-        );
-    }
-
-    #[test]
-    fn arm_pending_cancel_and_cancel_skips_cancel_on_stale_epoch() {
-        // reviewer 点 6 + 8：epoch 不匹配（轮次已切换）时，cancel 闭包不得
-        // 执行——stale 的取消不能落到新轮已 reset_cancel_token 的活跃 token 上。
-        let lifecycle = Arc::new(TurnLifecycle::default());
-        // turn1：on_submitted 激活（epoch=1）。
-        assert!(lifecycle.on_submitted());
-        let target = lifecycle.current_turn_generation().expect("turn1 epoch");
-
-        // 轮次切换：turn1 终态 → turn2 reserve + 提交（epoch=2）。
-        assert!(lifecycle.finish_once(|| {}).is_some());
-        let reservation2 = lifecycle.reserve().expect("turn2 reserve");
-        lifecycle.mark_reservation_submitted(reservation2.reservation_id);
-        let epoch2 = lifecycle
-            .current_turn_generation()
-            .expect("turn2 active epoch");
-        assert_eq!(epoch2, target + 1, "turn2 must bump the epoch past target");
-
-        // 用发起时快照 target 组合 arm+cancel：epoch 不匹配 → 返回 false 且
-        // cancel 闭包不执行（若误执行，探针会置位）。
-        let cancel_ran = Arc::new(AtomicBool::new(false));
-        let probe = cancel_ran.clone();
-        assert!(
-            !lifecycle.arm_pending_cancel_and_cancel(target, move || {
-                probe.store(true, Ordering::Release);
-            }),
-            "stale epoch must be rejected and the cancel closure skipped"
-        );
-        assert!(
-            !cancel_ran.load(Ordering::Acquire),
-            "cancel closure must not run when the epoch no longer matches"
-        );
-        assert!(
-            lifecycle.take_pending_cancel(epoch2).is_none(),
-            "stale cancel must not bind pending to the new turn"
-        );
-
-        // 防 Drop 副作用。
-        reservation2.mark_submitted();
-    }
-
-    #[test]
-    fn arm_pending_cancel_and_cancel_skips_cancel_when_idle_epoch_zero() {
-        // G2/epoch-0 哨兵守卫：fresh 会话（从未有过 turn，turn_epoch==0）时
-        // 发起 cancel，调用方以 target.unwrap_or(0) 传入 0，`turn_epoch != 0`
-        // 校验会通过（0 == 0）。若 engine 已为「即将启动的定时轮」在场
-        // （run_scheduled_turn 的 spawn→submit 窗口），执行 cancel 闭包会命中
-        // 该新 token，取消信号与 reset_cancel_token 竞速、结果不确定。
-        // idle 守卫必须跳过取消闭包（返回 true 但不执行），且不 arm pending。
-        let lifecycle = Arc::new(TurnLifecycle::default());
-        // fresh 会话：Default 的 TurnLifecycleState.turn_epoch == 0、active == false。
-        assert_eq!(
-            lifecycle.current_turn_generation(),
-            None,
-            "idle fresh session"
-        );
-
-        let cancel_ran = Arc::new(AtomicBool::new(false));
-        let probe = cancel_ran.clone();
-        // target=None 时调用方以 0 哨兵传入；epoch 匹配（0==0）但 idle →
-        // 不执行 cancel 闭包、返回 true（调用方按 armed 路径继续，只清遗留
-        // 子代理，不误伤尚未启动的定时轮）。
-        assert!(
-            lifecycle.arm_pending_cancel_and_cancel(0, move || {
-                probe.store(true, Ordering::Release);
-            }),
-            "idle must report success so the caller can continue (cascade only)"
-        );
-        assert!(
-            !cancel_ran.load(Ordering::Acquire),
-            "cancel closure must not run on an idle fresh session (epoch-0 sentinel)"
-        );
-        assert!(
-            lifecycle.take_pending_cancel(0).is_none(),
-            "idle must not arm pending_cancel"
-        );
-    }
-
-    #[test]
-    fn claim_unsubmitted_terminal_for_epoch_rejects_stale_target() {
-        // reviewer 点 7 的原子边界：未提交认领必须在 state 锁内与
-        // turn_epoch == target 同时校验。generation 检查与认领分离时，
-        // reserve_turn（不取 turn_lock）可在两者之间完成切轮，无 epoch 校验
-        // 的认领会把新轮 reservation 误认领为 Interrupted、清掉新轮 busy。
-        let lifecycle = Arc::new(TurnLifecycle::default());
-
-        // turn1：reserve 未提交（epoch=1），cancel 阶段二走认领路径。
-        let reservation1 = lifecycle.reserve().expect("turn1 reserve");
-        let target = lifecycle
-            .current_turn_generation()
-            .expect("turn1 active epoch");
-        assert_eq!(target, 1);
-
-        // 模拟「generation 检查通过后、认领前」另一 worker 切轮：
-        // turn1 终态（未提交认领路径）→ turn2 reserve（epoch=2，未提交）。
-        assert!(lifecycle.finish_unsubmitted_once());
-        let reservation2 = lifecycle.reserve().expect("turn2 reserve");
-        assert_eq!(
-            lifecycle.current_turn_generation(),
-            Some(target + 1),
-            "turn2 must bump the epoch past target"
-        );
-
-        // stale target 认领被拒：返回 false 且无副作用，turn2 完好。
-        assert!(
-            !lifecycle.claim_unsubmitted_terminal_for_epoch(target),
-            "claim with a stale target must no-op"
-        );
-        assert!(
-            reservation2.ensure_active().is_ok(),
-            "new turn unsubmitted reservation must survive a stale claim"
-        );
-
-        // 当前 epoch 认领成功：原子校验通过后正常认领，turn2 被失效。
-        assert!(lifecycle.claim_unsubmitted_terminal_for_epoch(target + 1));
-        assert!(
-            reservation2.ensure_active().is_err(),
-            "claimed reservation must be invalidated"
-        );
-        lifecycle.finish_terminal_emission();
-
-        // 防 Drop 副作用：turn1/turn2 均已认领终态（reservation 幂等）。
-        drop(reservation1);
-        drop(reservation2);
     }
 
     #[test]
@@ -3851,16 +1999,14 @@ mod turn_lifecycle_tests {
         let lifecycle = Arc::new(TurnLifecycle::default());
 
         lifecycle.on_submitted();
-        let epoch_prev = lifecycle.current_turn_generation().expect("active epoch");
-        lifecycle.arm_pending_cancel_and_cancel(epoch_prev, || {});
+        lifecycle.arm_pending_cancel();
         // 模拟 turn 未正常 started 就结束（如 engine spawn 失败）。
         assert!(lifecycle.finish_once(|| {}).is_some());
 
         // 新一轮 reserve 后，pending_cancel 必须被清除。
         let _reservation = lifecycle.reserve().expect("reserve");
-        let epoch_next = lifecycle.current_turn_generation().expect("active epoch");
         assert!(
-            lifecycle.take_pending_cancel(epoch_next).is_none(),
+            !lifecycle.take_pending_cancel(),
             "stale pending_cancel from previous turn must be cleared by reserve"
         );
     }
@@ -3875,96 +2021,20 @@ mod turn_lifecycle_tests {
         lifecycle.on_submitted();
 
         // cancel 路径：arm_pending_cancel（turn_id 仍为 None → 置标记）。
-        let epoch = lifecycle.current_turn_generation().expect("active epoch");
-        lifecycle.arm_pending_cancel_and_cancel(epoch, || {});
+        lifecycle.arm_pending_cancel();
 
         // Engine 执行 reset_cancel_token + 发 TurnStarted → 转发器先
         // on_started_transition（设 turn_id），再 take_pending_cancel。
-        // 同一轮内 on_started 不 bump epoch（active 已 true），take 仍匹配。
         lifecycle.on_started("turn-reset".to_string());
-        let pending = lifecycle.take_pending_cancel(epoch);
+        let pending = lifecycle.take_pending_cancel();
         assert!(
-            pending.is_some(),
+            pending,
             "pending_cancel must survive until TurnStarted consumes it"
         );
 
         // 消费后标记清除，下一轮不受影响。
-        assert!(lifecycle.take_pending_cancel(epoch).is_none());
+        assert!(!lifecycle.take_pending_cancel());
         assert!(lifecycle.finish_once(|| {}).is_some());
-    }
-
-    #[test]
-    fn turn_epoch_advances_on_every_admission_path_and_survives_terminal() {
-        // turn_epoch 是 cancel generation 的身份来源，必须在三个 active 入口
-        // （reserve / on_submitted / on_started_transition newly_active）都单调
-        // 自增，且与 is_active() 同口径：终态发送临界区（terminal_closing）内
-        // 仍是本轮 epoch，finish 后才回 None。
-        let lifecycle = Arc::new(TurnLifecycle::default());
-
-        // 空闲 → None。
-        assert_eq!(lifecycle.current_turn_generation(), None);
-
-        // reserve 推进到 epoch=1（reserve 后未 submitted，走未提交认领终态）。
-        let reservation = lifecycle.reserve().expect("reserve");
-        assert_eq!(lifecycle.current_turn_generation(), Some(1));
-
-        // 结束本轮（claim_unsubmitted 进入 terminal_closing 临界区，epoch 仍是 1）。
-        // claim 成功后 reservation 失效（ensure_active 失败），无需 mark_submitted。
-        assert!(lifecycle.claim_unsubmitted_terminal());
-        assert_eq!(
-            lifecycle.current_turn_generation(),
-            Some(1),
-            "terminal_closing 临界区内 generation 仍是本轮 epoch"
-        );
-        // finish 释放闸门后回到空闲 → None。
-        lifecycle.finish_terminal_emission();
-        drop(reservation);
-        assert_eq!(lifecycle.current_turn_generation(), None);
-
-        // finish 释放闸门后回到空闲 → None。
-        assert_eq!(lifecycle.current_turn_generation(), None);
-
-        // on_submitted（定时任务路径，非 reservation）推进到 epoch=2。
-        assert!(lifecycle.on_submitted());
-        assert_eq!(lifecycle.current_turn_generation(), Some(2));
-        assert!(lifecycle.finish_once(|| {}).is_some());
-
-        // on_started_transition 从 idle 激活（newly_active 分支）推进到 epoch=3。
-        assert!(lifecycle
-            .on_started_transition("turn-stale".to_string())
-            .is_some());
-        assert_eq!(lifecycle.current_turn_generation(), Some(3));
-        assert!(lifecycle.finish_once(|| {}).is_some());
-    }
-
-    #[test]
-    fn pending_cancel_tied_to_epoch_is_dropped_after_turn_change() {
-        // reviewer 点 4 的核心回归：并发取消请求 C2 在旧轮 arm 了 pending_cancel，
-        // 恢复时已变成新轮。pending_cancel 携带 epoch，take 时校验不匹配 → 丢弃，
-        // 不误取消新轮（不触发 approve_handle.cancel 重放）。
-        //
-        // 现有 pending_cancel_does_not_leak_across_turns 只测 reserve 清除，不覆盖
-        // 「新轮已 reserve 后 C2 才 arm」的窗口——此测试直接覆盖该缺口。
-        let lifecycle = Arc::new(TurnLifecycle::default());
-
-        // 旧轮：submit（epoch=1），arm pending_cancel 绑定到 epoch=1。
-        lifecycle.on_submitted();
-        let epoch_old = lifecycle.current_turn_generation().expect("old epoch");
-        lifecycle.arm_pending_cancel_and_cancel(epoch_old, || {});
-
-        // 结束旧轮，新一轮 reserve（epoch=2）。
-        assert!(lifecycle.finish_once(|| {}).is_some());
-        let _reservation = lifecycle.reserve().expect("new turn");
-        let epoch_new = lifecycle.current_turn_generation().expect("new epoch");
-        assert_ne!(epoch_old, epoch_new);
-
-        // forwarder 在新轮 TurnStarted 后用新轮 epoch 取 pending_cancel：
-        // C2 留下的 stale pending（epoch_old）必须被丢弃，不重放 cancel。
-        assert_eq!(
-            lifecycle.take_pending_cancel(epoch_new),
-            None,
-            "stale pending_cancel bound to a previous epoch must be dropped, not replayed onto the new turn"
-        );
     }
 
     #[test]
